+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased][unreleased]
 
+### Changed
+
+- Service support can view re-routing rules for "dual-running" of service with
+  the new .NET version
+
 ## [Release-115][release-115]
 
 ### Fixed